requirements.txt
argparse-addons==0.8.0
bitstruct==8.14.1
cantools==37.0.7
diskcache==5.4.0
msgpack==1.0.3
packaging==21.3
pyparsing==3.0.8
python-can==4.0.0
textparser==0.24.0
typing-extensions==4.2.0
wrapt==1.14.1
